requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements-dev.in
#
babel==2.18.0
    # via -r requirements-dev.in
build==1.4.2
    # via pip-tools
click==8.3.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
iniconfig==2.3.0
    # via pytest
packaging==26.0
    # via
    #   build
    #   pytest
    #   wheel
pip-tools==7.5.3
    # via -r requirements-dev.in
pluggy==1.6.0
    # via pytest
pygments==2.20.0
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.2
    # via -r requirements-dev.in
python-dotenv==1.2.2
    # via -r requirements.in
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
